Create DebitCardReward.sol
Here is a Solidity contract that tracks the number of times an address has used their debit card and rewards them with an ERC20 token when their usage is above 50 times per month. This contract keeps track of the number of times each debit card holder uses their card, and when a card holder reaches 50 uses in a month, it rewards them with the specified amount of ERC20 tokens. 
This contract has a trackUsage function that increments the usage count for an address every time it is called. If the usage count for an address reaches 50 or a multiple of 50, the contract rewards the address with an ERC20 token. The contract also has a resetUsage function that can be called by the owner to reset the usage count for all addresses to 0 at the end of each month.