Create DebitCardReward.js
This code can be used in a front-end application, such as a web page, to allow users to interact with the contract and track their debit card usage. The trackUsage function can be called every time the user makes a purchase with their debit card, and the resetUsage function can be called at the end of each month to reset the usage count for all users. The ThresholdReached event can be used to notify users when they have reached the usage threshold and are eligible for a reward.